Remove newMain.cpp file containing order management logic and related functionality.
Moved DemoMain.cpp logic into integrated_main.cpp to showcase both greenhouse and sales-floor fucntionality in one test.

+++ SuggestionTemplate/BouquetSuggestion.h
@@ -32,7 +32,7 @@ struct BouquetSuggestion {
         oss << "   Colors: " << colorScheme << "\n";
         oss << "   Flowers: ";
         
-        for (int i = 0; i < flowerTypes.size(); i++) {
+        for (size_t i = 0; i < flowerTypes.size(); i++) {
             oss << quantities[i] << "x " << flowerTypes[i];
             if (i < flowerTypes.size() - 1) oss << ", ";
         }
